google_oauth-20
This is a MongoDB authentication system by passport library strategy google_oauth-20. In this work, I made a sign-up and in the system through a google account. Firstly, I created an account at google developers console and then created a project for generating ClientId and ClientSecret to connect my web application with the cloud. In this google strategy, first I found the google id as my profile id and then took the name of the user as my profile name. Then, I created the session id that whenever we log in creates a user id inside the session. Finally, I did serialize and deserialized it for the user. In this way, the user does not need to register. it's directly authenticated by google id which is also an advanced technology to serve.